Let openstack driver work against https or http (APPENG-3661)
Add a method, _secureToInsecureFallback, that takes a callback method,
args and kwargs.   That method adds secure=True to the kwargs and calls
the callback in a try.  If it fails, it adds secure=False to the kwargs
and calls the callback again, in a separate try.  If both fail, it
raises the secure=True exception so that users seeing the URL won't be
confused because the secure=False test happened last.  The callback
method is responsible for handling the 'secure' keyword to do whatever
behavior change is necessary.

Move the authentication and the server connection test code into
callbacks that are then called by _secureToInsecureFallback.

Changing the test code to work against both https and http would be very
hard, since there are lots of hardcoded URLs in the data that the REST
server uses.  Instead, just test https in the bulk of the unit tests,
but add a test of the fallback logic.